ui: move tutorial_dodge_hint.h from app/ui/ to app/util/ (sub-PR 2 of 3)
Per #1317 (§ 7 folder allowlist): `app/ui/tutorial_dodge_hint.h` is a
header-only `inline` helper consumed by `tutorial_dodge_hint_bind_system`
plus two test TUs. It is not UI — it is a tiny runtime input-capability
selector for the Tutorial dodge hint (#513). Moving it to `app/util/`
fits Section-7's 5-folder allowlist and lines up sub-PR 3 (the final
`generated/*` move) to be the only thing left between us and a clean
`app/` layout.

Mechanical changes:

- `git mv app/ui/tutorial_dodge_hint.h app/util/tutorial_dodge_hint.h`
  (header guard `SHAPESHIFTER_TUTORIAL_DODGE_HINT_H` is path-agnostic).
- Update the 3 include sites:
  - `app/systems/tutorial_dodge_hint_bind_system.cpp`
  - `tests/test_ui_update_system.cpp`
  - `tests/test_tutorial_dodge_hint.cpp`
- Update `design-docs/rguilayout-portable-c-integration.md`:
  - Move the helper out of the `app/ui/` directory diagram into the
    `app/util/` diagram.
  - Resolve open decision #1 to "moved to app/util/" with citation.

No CMake change is needed: `UI_SOURCES` is `app/ui/*.cpp` only, and the
helper is a header; the existing `app/util/*.cpp` glob already picks up
util sources (this header is not compiled directly).

After sub-PR 1 (#1321) the only contents of `app/ui/` were this header
plus `generated/`. Once sub-PR 3 lands, `app/ui/` disappears entirely
and `app/` matches the Section-7 allowlist exactly.

Refs #1317, #1193
